Rename Prebid config file.
