+++ web/src/js/ads/__tests__/ads.test.js
@@ -8,7 +8,7 @@ const mockGoogleAdSlotDefinitions = jest.fn()
 const mockOpenXconfig = jest.fn()
 jest.mock('../adsEnabledStatus', () => mockAdsEnabledStatus)
 jest.mock('../prebid/prebidModule', () => mockPrebidModule)
-jest.mock('../prebid/config', () => mockPrebidConfig)
+jest.mock('../prebid/prebidConfig', () => mockPrebidConfig)
 jest.mock('../googleTagManager', () => mockGoogleTagManager)
 jest.mock('../googleAdSlotDefinitions', () => mockGoogleAdSlotDefinitions)
 jest.mock('../openxConfig', () => mockOpenXconfig)
